cgame: WIP new hud editor layout
